Account for relative shape size. Improve accuracy
Update the matching of shape parts to take into consideration the
relative size of each part, with respect to its entire shape.  This
avoids very small shape parts matching against very large shape parts in
another model shape.  Also, we now avoid leaving large shape parts
unmatched, since that is unintuitive and accounts for some of the errors
we saw earlier.   This still needs a bit of work, but we do see some
improvement.

Finally, updated .gitignore with a mass .gitignore list found on
stackoverflow.  Looks good, and reminded me that .suo files don't need
to be in version control.