Implemented matching with the learned parsing model.
Finished an implementation of matching with the learned parsing model.
I have kept it entirely separate from the previous matching algorithm,
because we're not where we want to be with it yet.  There is a problem
with the learning of the model that is causing poor accuracy.
Basically, the learning model is going on a class by class basis and
learning the MAP parameterization of that class.  What we _really_ want
is the _joint_ MAP parameterization over all classes.  This is very very
different, and much more computationally costly.  Need to think about
that some more.